+++ RELEASE_NOTES.md
@@ -58,6 +58,7 @@ A `> v8.0: Comment` block will be added to the docs, where structural or archite
 
 * `AutocompleteCompletion` has been deprecated.
 * `AutocompleteResult` has been deprecated.
+* `AutocompleteSpaceState` has been deprecated.
 * `AutocompleteSuggestion.isAutocomplete` has been renamed to `isAutocorrect`.
 * `AutocompleteToolbarStyle` has been renamed to `KeyboardStyle.AutocompleteToolbarItem`.
 * `AutocompleteToolbarItemStyle` has been renamed to `KeyboardStyle.AutocompleteToolbarItem`.